Created an index page file to serve as the main landing page.
Created an index landing page to display the main recipes. When the user starts up the project,
they will start on the index page and they can gradually browse through the recipes using the links.
The first link for the first recipe was added and redirects to the relative file inside the recipes folder.